requirements.txt
flask
scikit-learn
python-dotenv
numpy
google-genai